--- src/main/java/com/gdgoc/Online_Library/Common/Role.java
@@ -1,5 +1,0 @@
-package com.gdgoc.Online_Library.Common;
-
-public enum Role {
-    ADMIN, USER
-}
