+++ src/main/java/com/gdgoc/Online_Library/book/BookController.java
@@ -1,8 +1,8 @@
-package com.gdgoc.Online_Library.Book;
+package com.gdgoc.Online_Library.book;
 
 import static org.springframework.http.HttpStatus.OK;
 
-import com.gdgoc.Online_Library.Book.dto.BookGetResponse;
+import com.gdgoc.Online_Library.book.dto.BookGetResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
