+++ src/main/java/com/gdgoc/Online_Library/common/Role.java
@@ -1,0 +1,5 @@
+package com.gdgoc.Online_Library.common;
+
+public enum Role {
+    ADMIN, USER
+}
