+++ src/main/java/com/gdgoc/Online_Library/common/Status.java
@@ -1,4 +1,4 @@
-package com.gdgoc.Online_Library.Common;
+package com.gdgoc.Online_Library.common;
 
 public enum Status {
     AVAILABLE, BORROWED, LOST
